HERCULES-8873 - [UCV][SGE] Proyectos - Amortización fondos - Configuración integración

+++ sgi-webapp/src/app/module/csp/proyecto/proyecto-formulario/proyecto-amortizacion-fondos/proyecto-amortizacion-fondos.component.ts
@@ -20,11 +20,13 @@ import { PROYECTO_ROUTE_NAMES } from '../../proyecto-route-names';
 import { ProyectoActionService } from '../../proyecto.action.service';
 import { IEntidadFinanciadora } from '../proyecto-entidades-financiadoras/proyecto-entidades-financiadoras.fragment';
 import { IProyectoPeriodoAmortizacionListado, ProyectoAmortizacionFondosFragment } from './proyecto-amortizacion-fondos.fragment';
+import { ConfigService } from '@core/services/csp/config.service';
 
 const MSG_NUEVO = marker('title.new.entity');
 const MSG_DELETE = marker('msg.delete.entity');
 const PROYECTO_ENTIDAD_FINANCIADORA_KEY = marker('csp.proyecto-amortizacion-fondos.entidad-financiadora');
 const PROYECTO_PERIODO_AMORTIZACION_KEY = marker('csp.proyecto-amortizacion-fondos.periodo-amortizacion');
+const MSG_DELETE_AMORTIZACION_FONDOS_DISABLED = marker('msg.sge.delete-amortizacion-fondos-disabled');
 
 @Component({
   selector: 'sgi-proyecto-amortizacion-fondos',
@@ -68,11 +70,17 @@ export class ProyectoAmortizacionFondosComponent extends FragmentComponent imple
   @ViewChild('paginatorPeriodos', { static: true }) paginatorPeriodos: MatPaginator;
   @ViewChild('sortPeriodos', { static: true }) sortPeriodos: MatSort;
 
+  private _amortizacionFondosSgeEnabled: boolean;
+  get amortizacionFondosSgeDisabled(): boolean {
+    return !this._amortizacionFondosSgeEnabled;
+  }
+
   constructor(
     protected actionService: ProyectoActionService,
     private matDialog: MatDialog,
     private dialogService: DialogService,
     private readonly translate: TranslateService,
+    private readonly cspConfigService: ConfigService
   ) {
     super(actionService.FRAGMENT.AMORTIZACION_FONDOS, actionService);
     this.formPart = this.fragment as ProyectoAmortizacionFondosFragment;
@@ -87,6 +95,12 @@ export class ProyectoAmortizacionFondosComponent extends FragmentComponent imple
     this.fxLayoutProperties.gap = '20px';
     this.fxLayoutProperties.layout = 'row wrap';
     this.fxLayoutProperties.xs = 'column';
+
+    this.subscriptions.push(
+      this.cspConfigService.isAmortizacionFondosSgeEnabled().subscribe(amortizacionFondosSgeEnabled => {
+        this._amortizacionFondosSgeEnabled = amortizacionFondosSgeEnabled;
+      })
+    );
   }
 
   ngOnInit(): void {
@@ -154,7 +168,8 @@ export class ProyectoAmortizacionFondosComponent extends FragmentComponent imple
       proyectoId: this.formPart.getKey() as number,
       entidadesFinanciadoras: this.formPart.entidadesFinanciadoras$.value.map(entidad => entidad.value),
       proyectosSGE: this.formPart.proyectosSGE$.value,
-      anualidadGenerica: !this.formPart.anualidades
+      anualidadGenerica: !this.formPart.anualidades,
+      amortizacionFondosSgeDisabled: this.amortizacionFondosSgeDisabled
     };
     const config = {
       data
@@ -175,16 +190,21 @@ export class ProyectoAmortizacionFondosComponent extends FragmentComponent imple
 
 
   deletePeriodoAmortizacion(wrapper: StatusWrapper<IProyectoPeriodoAmortizacion>) {
-    this.subscriptions.push(
-      this.dialogService.showConfirmation(this.textoDeletePeriodosAmortizacion).subscribe(
-        (aceptado) => {
-          if (aceptado) {
-            this.formPart.deletePeriodoAmortizacion(wrapper);
-            this.formPart.recalcularNumPeriodos();
+    if (this.amortizacionFondosSgeDisabled) {
+      this.subscriptions.push(
+        this.dialogService.showInfoDialog(MSG_DELETE_AMORTIZACION_FONDOS_DISABLED).subscribe());
+    } else {
+      this.subscriptions.push(
+        this.dialogService.showConfirmation(this.textoDeletePeriodosAmortizacion).subscribe(
+          (aceptado) => {
+            if (aceptado) {
+              this.formPart.deletePeriodoAmortizacion(wrapper);
+              this.formPart.recalcularNumPeriodos();
+            }
           }
-        }
-      )
-    );
+        )
+      );
+    }
   }
 
   ngOnDestroy(): void {
